requirements.txt
torch
torchvision
pyarrow
pillow
tf-yarn
wandb==0.12.11
webdataset>=0.2.5
nltk
fasttext==0.9.1
easynmt
ml-hadoop-experiment
thx
fsspec
multilingual-clip
open-clip-torch-any-py3
